one step to be conform to ISO C + headers LGPL added to exports.h and basic_types.h
git-svn-id: http://beaengine.googlecode.com/svn/trunk@38 e67b0f8a-f93e-11de-ba51-6b4196a2b81e

+++ beaengineSources/BeaEngine.h
@@ -68,7 +68,7 @@ typedef struct {
 #pragma pack(1)
 typedef struct  {
    long Category;
-   UInt64 Opcode;
+   long Opcode;
    char Mnemonic[16];
    long BranchType;
    EFLStruct Flags;
@@ -290,6 +290,6 @@ enum SPECIAL_INFO
 #ifdef __cplusplus
 extern "C"
 #endif
-Int64 __bea_callspec__ Disasm(LPDISASM pDisAsm);
+size_t __bea_callspec__ Disasm(LPDISASM pDisAsm);
 
 #endif
